function that returns the number of bits you would need to flip

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,8 +9,8 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int k = 0;
-	unsigned int xor = n ^ m;
+	unsigned long int k = 0;
+	unsigned long int xor = n ^ m;
 
 	while (xor)
 	{
